Fix bug in AdjSpreadCountReducer

+++ cosmos/models/mobility/src/main/java/es/tid/cosmos/mobility/adjacentextraction/AdjSpreadCountReducer.java
@@ -6,7 +6,7 @@ import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import es.tid.cosmos.base.data.TypedProtobufWritable;
-import es.tid.cosmos.base.data.generated.BaseTypes.Int;
+import es.tid.cosmos.base.data.generated.BaseTypes.Int64;
 import es.tid.cosmos.mobility.data.generated.MobProtocol.TwoInt;
 
 /**
@@ -15,13 +15,14 @@ import es.tid.cosmos.mobility.data.generated.MobProtocol.TwoInt;
  * 
  * @author dmicol
  */
-class AdjSpreadCountReducer extends Reducer<LongWritable,
-        TypedProtobufWritable<TwoInt>, LongWritable, TypedProtobufWritable<Int>> {
+class AdjSpreadCountReducer extends Reducer<
+        LongWritable, TypedProtobufWritable<TwoInt>,
+        LongWritable, TypedProtobufWritable<Int64>> {
     @Override
     protected void reduce(LongWritable key,
             Iterable<TypedProtobufWritable<TwoInt>> values, Context context)
             throws IOException, InterruptedException {
-        int valueCount = 0;
+        long valueCount = 0L;
         for (TypedProtobufWritable<TwoInt> value : values) {
             valueCount++;
         }
